+++ client/src/extension.js
@@ -35,6 +35,7 @@ function activate(context) {
 		}
 	}
 	var cl = new client.LanguageClient('HarbourServer', 'Harbour Server', serverOptions, clientOptions);
+	cl.registerProposedFeatures()
 	context.subscriptions.push(cl.start());
 	vscode.commands.registerCommand('harbour.getdbgcode', GetDbgCode);
 	vscode.commands.registerCommand("harbour.debugList", DebugList)